Decouple project links from main layout
